requirements.txt
openai, 
pandas, 
numpy, 
scikit-learn, 
hdbscan,
umap-learn,
matplotlib, 
streamlit, 
tqdm.